// src/payload-types.ts
export interface Media {
  id: number;
  alt: string;
  category?: ('brand' | 'logo' | 'favicon' | 'project' | 'post' | 'other') | null;
  tags?: string[] | null;
  prefix?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
// ...
export interface Profile {
  id: number;
  fullName: string;
  /**
   * A brief 1-2 sentence introduction for the hero section.
   */
  bioShort: string;
  location?: string | null;
  profileImage?: (number | null) | Media;
  resume?: (number | null) | Media;
  disciplines?:
    | {
        name: string;
        id?: string | null;
      }[]
    | null;
  contactEmail?: string | null;
  socialLinks?:
    | {
        platform: 'twitter' | 'github' | 'linkedin' | 'dribbble' | 'instagram' | 'youtube' | 'facebook' | 'website';
        url: string;
        id?: string | null;
      }[]
    | null;
  updatedAt?: string | null;
  createdAt?: string | null;
}